+++ ocaml-lsp-server/src/custom_requests/custom_request.mli
@@ -16,3 +16,4 @@ module Phrase = Req_phrase
 module Type_expression = Req_type_expression
 module Locate = Req_locate
 module Destruct = Req_destruct
+module Locate_types = Req_locate_types